Restructuration fichier decor
