--- FinalProjet/FinalProject/Debug/FinalProject.Build.CppClean.log
@@ -1,7 +1,0 @@
-S:\2015\VISION\LeapMotion\FinalProject\FinalProject\Debug\cl.command.1.tlog
-S:\2015\VISION\LeapMotion\FinalProject\FinalProject\Debug\CL.read.1.tlog
-S:\2015\VISION\LeapMotion\FinalProject\FinalProject\Debug\CL.write.1.tlog
-S:\2015\VISION\LeapMotion\FinalProject\FinalProject\Debug\FinalProject.write.1.tlog
-S:\2015\VISION\LEAPMOTION\FINALPROJECT\FINALPROJECT\DEBUG\LEAPLISTENER.OBJ
-S:\2015\VISION\LeapMotion\FinalProject\FinalProject\Debug\vc100.idb
-S:\2015\VISION\LEAPMOTION\FINALPROJECT\FINALPROJECT\DEBUG\VC100.PDB
