+++ FinalProjet/FinalProject/Debug/FinalProject.lastbuildstate
@@ -1,2 +1,2 @@
-#v4.0:v100:false
-Debug|Win32|C:\Users\Usager\Desktop\FinalProjet\|
+#v4.0:v100
+Debug|Win32|S:\2015\VISION\LeapMotion\GIT\h15-monkey-thug-life\FinalProjet\|
