+++ FinalProjet/FinalProject/Debug/FinalProject.log
@@ -1,17 +1,7 @@
-﻿La génération a démarré 2015-03-03 11:39:34.
-     1>Projet "C:\Users\Usager\Desktop\FinalProjet\FinalProject\FinalProject.vcxproj" sur le nœud 2 (build cible(s)).
-     1>InitializeBuildStatus:
-         Création de "Debug\FinalProject.unsuccessfulbuild", car "AlwaysCreate" a été spécifié.
-       ClCompile:
-         Toutes les sorties sont à jour.
-       Link:
-         Toutes les sorties sont à jour.
-         FinalProject.vcxproj -> C:\Users\Usager\Desktop\FinalProjet\Debug\FinalProject.exe
-       FinalizeBuildStatus:
-         Suppression du fichier "Debug\FinalProject.unsuccessfulbuild".
-         Mise à jour de l'horodatage "Debug\FinalProject.lastbuildstate".
-     1>Génération du projet "C:\Users\Usager\Desktop\FinalProjet\FinalProject\FinalProject.vcxproj" terminée (build cible(s)).
-
-La génération a réussi.
-
-Temps écoulé 00:00:00.15
+﻿  Arbre.cpp
+s:\2015\vision\leapmotion\git\h15-monkey-thug-life\finalprojet\finalproject\gestiondecor\arbre.cpp(66): warning C4244: '-=' : conversion de 'float' en 'int', perte possible de données
+  Génération de code en cours...
+  Compilation en cours...
+  main.cpp
+  Génération de code en cours...
+  FinalProject.vcxproj -> S:\2015\VISION\LeapMotion\GIT\h15-monkey-thug-life\FinalProjet\Debug\FinalProject.exe
