+++ FinalProjet/FinalProject/FinalProject.vcxproj
@@ -68,32 +68,32 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClCompile Include="Arbre.cpp" />
     <ClCompile Include="Fonctionnalite\BaseFunc.cpp" />
     <ClCompile Include="Fonctionnalite\Ressource.cpp" />
     <ClCompile Include="GestionAnimation\Animation.cpp" />
     <ClCompile Include="GestionAnimation\AnimationPlayer.cpp" />
+    <ClCompile Include="GestionDecor\Arbre.cpp" />
+    <ClCompile Include="GestionDecor\Nuage.cpp" />
     <ClCompile Include="GestionRoute\Bordure.cpp" />
     <ClCompile Include="GestionRoute\Route.cpp" />
     <ClCompile Include="GestionRoute\Voie.cpp" />
     <ClCompile Include="LeapMotion\LeapListener.cpp" />
     <ClCompile Include="main.cpp" />
-    <ClCompile Include="Nuage.cpp" />
     <ClCompile Include="Perso\Monkey.cpp" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="Arbre.h" />
     <ClInclude Include="Fonctionnalite\BaseFunc.h" />
     <ClInclude Include="Fonctionnalite\Ressource.h" />
     <ClInclude Include="GestionAnimation\Animation.h" />
     <ClInclude Include="GestionAnimation\AnimationPlayer.h" />
+    <ClInclude Include="GestionDecor\Arbre.h" />
+    <ClInclude Include="GestionDecor\Nuage.h" />
     <ClInclude Include="GestionRoute\Bordure.h" />
     <ClInclude Include="GestionRoute\Route.h" />
     <ClInclude Include="GestionRoute\Voie.h" />
     <ClInclude Include="include\Leap.h" />
     <ClInclude Include="include\LeapMath.h" />
     <ClInclude Include="LeapMotion\LeapListener.h" />
-    <ClInclude Include="Nuage.h" />
     <ClInclude Include="Perso\Monkey.h" />
     <ClInclude Include="SDL\Image\SDL_image.h" />
     <ClInclude Include="SDL\include\begin_code.h" />
