+++ FinalProjet/FinalProject/GestionDecor/Arbre.h
@@ -1,0 +1,25 @@
+#pragma once
+#include "..\Fonctionnalite\BaseFunc.h"
+#include "..\Fonctionnalite\Ressource.h"
+
+class Arbre
+{
+public:
+	Arbre(void);
+	~Arbre(void);
+	Arbre(int Positionx, int Positiony,float Speed, Ressource Ressources, int Mode);
+	SDL_Surface* TabArbreTexture; 
+	int Positionx, Positiony;
+	SDL_Rect* RecDestination;
+	SDL_Rect* RecDestination2;
+	SDL_Rect* RecDestination3;
+	SDL_Rect* RecDestination4;
+	SDL_Rect* RecDestination5;
+	SDL_Rect* RecDestination6;
+	float Speed;
+	int Mode;
+
+	void Update();
+	void Draw(SDL_Surface* &SurfaceDessin);
+};
+
